+++ mop-domain/src/entities/lenders/ILender.ts
@@ -1,10 +1,9 @@
-import {Loan} from "../loan";
+import {ILoan} from "../loan";
 import {IThing} from "../thing"
 
 export interface ILender {
     readonly id: string
     readonly items: Iterable<IThing>
-    acceptReturn(loan: Loan): Loan
-
-    markReturnAsDamaged(loan: Loan): Loan
+    startReturn(loan: ILoan): ILoan
+    finishReturn(loan: ILoan): ILoan
 }
