--- mop-domain/src/entities/lenders/externalLibrary.d.ts
@@ -1,7 +1,0 @@
-import { Loan } from "../loan";
-import { ILender } from "./ILender";
-import { ILibrary } from "../library";
-export declare class ExternalLibrary implements ILender, ILibrary {
-    acceptReturn(loan: Loan): Loan;
-    markReturnAsDamaged(loan: Loan): Loan;
-}
