--- mop-domain/src/entities/lenders/externalLibrary.js
@@ -1,13 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.ExternalLibrary = void 0;
-class ExternalLibrary {
-    acceptReturn(loan) {
-        return undefined;
-    }
-    markReturnAsDamaged(loan) {
-        return undefined;
-    }
-}
-exports.ExternalLibrary = ExternalLibrary;
-//# sourceMappingURL=externalLibrary.js.map
