--- mop-domain/src/entities/lenders/individualLender.ts
@@ -1,27 +1,0 @@
-import {Loan} from "../loan";
-import {Person} from "../person";
-import {LoanStatus} from "../../valueItems/loanStatus";
-import {ILender} from "./ILender";
-
-export class IndividualLender extends Person implements ILender{
-    acceptReturn(loan: Loan): Loan {
-        return new Loan(
-            loan.item,
-            loan.borrower,
-            loan.dueDate,
-            LoanStatus.RETURNED,
-            loan.returnLocation
-        )
-    }
-
-    markReturnAsDamaged(loan: Loan): Loan {
-        return new Loan(
-            loan.item,
-            loan.borrower,
-            loan.dueDate,
-            LoanStatus.RETURNED_DAMAGED,
-            loan.returnLocation
-        )
-    }
-
-}
