+++ mop-domain/src/entities/libraries/ILibrary.ts
@@ -1,0 +1,17 @@
+import {IBorrower} from "../borrower";
+import {IThing} from "../thing";
+import {ILoan} from "../loan";
+
+export interface ILibrary {
+    readonly name: string
+    readonly chargesFees: boolean
+    readonly borrowers: Iterable<IBorrower>
+    readonly availableItems: Iterable<IThing>
+    readonly allItems: Iterable<IThing>
+
+    canBorrow(borrower: IBorrower): boolean
+
+    borrow(item: IThing, borrower: IBorrower, until: Date): ILoan
+
+    return(loan: ILoan): ILoan
+}
