+++ mop-domain/src/repositories/libraryRepository.ts
@@ -1,6 +1,6 @@
-import {ILibrary} from "../entities/library";
 import {IRepository} from "./IRepository";
 import {User} from "../entities/user";
+import {ILibrary} from "../entities/libraries/ILibrary";
 
 
 export interface ILibraryRepository extends IRepository<ILibrary>{
